requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
assemblyai==0.20.0
async-timeout==4.0.3
atlassian-python-api==3.41.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
Brotli==1.1.0
bs4==0.0.2
build==1.1.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
ci-info==0.3.0
click==8.1.7
coloredlogs==15.0.1
configobj==5.0.8
configparser==6.0.0
cryptography==41.0.7
dataclasses-json==0.6.3
Deprecated==1.2.14
deprecation==2.1.0
dirtyjson==1.0.8
distro==1.8.0
dnspython==2.4.2
etelemetry==0.3.1
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.1
fitz==0.0.1.dev2
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-secret-manager==2.16.4
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpc-google-iam-v1==0.12.6
grpcio==1.59.2
grpcio-status==1.59.2
gunicorn==21.2.0
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.1.2
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kubernetes==29.0.0
langchain==0.0.341
langchain-core==0.0.6
langsmith==0.0.67
llama-hub==0.0.43
llama-index==0.10.15
llama-index-agent-openai==0.1.5
llama-index-cli==0.1.7
llama-index-core==0.10.15
llama-index-embeddings-openai==0.1.6
llama-index-indices-managed-llama-cloud==0.1.3
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.7
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.4
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.6
llama-index-readers-llama-parse==0.1.3
llama-index-vector-stores-chroma==0.1.5
llama-parse==0.3.5
llamaindex-py-client==0.1.13
loguru==0.7.2
looseversion==1.3.0
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mutagen==1.47.0
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nibabel==5.1.0
nipype==1.8.6
nltk==3.8.1
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.13.3
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.1.3
pathlib==1.0.1
pdfminer==20191125
pdfminer.six==20231228
pikepdf==8.6.0
Pillow==10.1.0
pinecone-client==2.2.4
posthog==3.5.0
proto-plus==1.22.3
protobuf==4.25.0
prov==2.0.0
psutil==5.9.6
pulsar-client==3.4.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
pydot==1.4.2
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyparsing==3.1.1
pypdf==4.1.0
PyPika==0.48.9
pyproject_hooks==1.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyxnat==1.6
PyYAML==6.0.1
rdflib==7.0.0
regex==2023.10.3
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
retrying==1.3.4
rsa==4.9
safetensors==0.4.2
scipy==1.11.3
selenium==4.8.3
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.1
tldextract==5.1.1
tokenizers==0.15.1
tomli==2.0.1
tqdm==4.66.1
traits==6.3.2
transformers==4.37.2
trio==0.24.0
trio-websocket==0.11.1
triton==2.1.0
tweepy==4.14.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.3
yt-dlp==2023.11.16
zipp==3.17.0
